Set up view layout, initalised the AuthService, set up models to handle response
Added proper HTML elements and Html form to the Authentication views. Used bootstrap for styling the views.
Added a directory designated to store the auth models called AuthModels. Created SignInModel, which is the object the for sends to the Signin method part of the AuthenticationController.  SignupModel is respectively the model that is utilised for the same purpose by the Signup method part of the Authentication controller. Added also the SigninResponseObject, which is the model that handles the response object, when it's being deserilized after fetched by the AuthService.
Registered the AuthService as a service in the Program.cs file.