Changed the way IAuthService is injected in the controllers and added error handling
There was a mistake in how the AuthService was injected as a service, I fixed it and replaced AuthService with IAuthService.
Added error handling in the AuthSignin method, part of the Authentication controller. Now the method handles exceptions and redirects accrording to the result of the request, sent to the Signin route on the noeTaskManagerService server.