Fixed Json Deserialization issues
There was a problem with the way the Json object received by the SignIn method was deserialized. Instead of access_token, as the server sends, the SigningResponseObject was expecting AccessToken or accesstoken. The secound issue was from the same kind, but with the UserResponseCreds object.
Fixed thos issues and now the SignIn flow works smoothly.